feat: add useJoeEntites reusable hooks, refactor joe useBestTrade and related components
