+++ src/components/Web3Status/index.tsx
@@ -7,8 +7,8 @@ import WalletDropdown from 'components/WalletDropdown'
 import { getConnection } from 'connection/utils'
 import { SupportedChainId } from 'constants/chains'
 import { NavBarVariant, useNavBarFlag } from 'featureFlags/flags/navBar'
+import { useJoeBestTrade } from 'hooks/avalanche/useJoeBestTrade'
 import { useBestTrade } from 'hooks/useBestTrade'
-import { useBestTradeByJoe } from 'hooks/useBestTradeTraderJoe'
 import { Portal } from 'nft/components/common/Portal'
 import { getIsValidSwapQuote } from 'pages/Swap'
 import { darken } from 'polished'
@@ -83,7 +83,7 @@ const Web3StatusConnectButton = styled.button<{ faded?: boolean }>`
   }
 `
 
-const Web3StatusConnect = styled(Web3StatusGeneric)<{ faded?: boolean }>`
+const Web3StatusConnect = styled(Web3StatusGeneric) <{ faded?: boolean }>`
   background-color: ${({ theme }) => theme.deprecated_primary4};
   border: none;
   color: ${({ theme }) => theme.deprecated_primaryText1};
@@ -110,7 +110,7 @@ const Web3StatusConnect = styled(Web3StatusGeneric)<{ faded?: boolean }>`
     `}
 `
 
-const Web3StatusConnected = styled(Web3StatusGeneric)<{ pending?: boolean }>`
+const Web3StatusConnected = styled(Web3StatusGeneric) <{ pending?: boolean }>`
   background-color: ${({ pending, theme }) => (pending ? theme.deprecated_primary1 : theme.deprecated_bg1)};
   border: 1px solid ${({ pending, theme }) => (pending ? theme.deprecated_primary1 : theme.deprecated_bg1)};
   color: ${({ pending, theme }) => (pending ? theme.deprecated_white : theme.deprecated_text1)};
@@ -122,7 +122,7 @@ const Web3StatusConnected = styled(Web3StatusGeneric)<{ pending?: boolean }>`
     :focus {
       border: 1px solid
         ${({ pending, theme }) =>
-          pending ? darken(0.1, theme.deprecated_primary1) : darken(0.1, theme.deprecated_bg2)};
+    pending ? darken(0.1, theme.deprecated_primary1) : darken(0.1, theme.deprecated_bg2)};
     }
   }
 `
@@ -174,10 +174,10 @@ const StyledConnect = styled.div`
   &:hover {
     color: ${({ theme }) => theme.accentActionSoft};
     transition: ${({
-      theme: {
-        transition: { duration, timing },
-      },
-    }) => `${duration.fast} color ${timing.in}`};
+  theme: {
+    transition: { duration, timing },
+  },
+}) => `${duration.fast} color ${timing.in}`};
   }
 `
 
@@ -191,7 +191,7 @@ function Web3StatusInnerDefault() {
 }
 
 function Web3StatusInnerAvalanche() {
-  return <BaseWeb3StatusInner useBestTradeHook={useBestTradeByJoe} />
+  return <BaseWeb3StatusInner useBestTradeHook={useJoeBestTrade} />
 }
 
 function Web3StatusInner() {
